camera: move Camera_ResetPosition to TRX
