+++ src/libtrx/include/libtrx/game/camera/const.h
@@ -2,4 +2,5 @@
 
 #include "../const.h"
 
+#define NO_CAMERA (-1)
 #define CAMERA_DEFAULT_DISTANCE (WALL_L * 3 / 2)
