+++ src/tr1/game/camera/common.h
@@ -5,6 +5,5 @@
 #include <stdint.h>
 
 void Camera_Initialise(void);
-void Camera_ResetPosition(void);
 void Camera_UpdateCutscene(void);
 void Camera_MoveManual(void);
