+++ src/tr2/game/camera.h
@@ -5,7 +5,6 @@
 #include <libtrx/game/camera.h>
 
 void Camera_Initialise(void);
-void Camera_ResetPosition(void);
 void Camera_Clip(
     int32_t *x, int32_t *y, int32_t *h, int32_t target_x, int32_t target_y,
     int32_t target_h, int32_t left, int32_t top, int32_t right, int32_t bottom);
